0206-2

+++ push_swap/libft/ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: gkwon <gkwon@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/16 11:06:27 by gkwon             #+#    #+#             */
-/*   Updated: 2022/11/20 21:50:35 by gkwon            ###   ########.fr       */
+/*   Updated: 2023/02/06 21:07:06 by gkwon            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ static int	long_long_return(int flag)
 	return (-1);
 }
 
-int	ft_atoi(const char *str)
+long	ft_atoi(const char *str)
 {
 	long long	ret;
 	long long	tmp;
